+++ internal/useragent/useragent.go
@@ -19,6 +19,7 @@ const (
 
 // settable by tests
 var goos = runtime.GOOS
+var executable = os.Executable
 
 func Get() string {
 	parenthesis := ""
@@ -27,7 +28,7 @@ func Get() string {
 		metadata = append(metadata, mode)
 	}
 	metadata = append(metadata, goos)
-	if op := getDeployMode(); op != "" {
+	if op := GetDeployMode(); op != "" {
 		metadata = append(metadata, op)
 	}
 	if len(metadata) > 0 {
@@ -49,12 +50,18 @@ func getRunMode() string {
 	}
 }
 
-func getDeployMode() string {
+// GetDeployMode returns our best-effort guess at the way Grafana Agent was deployed.
+func GetDeployMode() string {
 	op := os.Getenv(deployModeEnv)
 	// only return known modes. Use "binary" as a default catch-all.
 	switch op {
 	case "operator", "helm", "docker", "deb", "rpm", "brew":
 		return op
 	}
+	// try to detect if executable is in homebrew directory
+	if path, err := executable(); err == nil && goos == "darwin" && strings.Contains(path, "brew") {
+		return "brew"
+	}
+	// fallback to binary
 	return "binary"
 }