+++ pkg/usagestats/stats.go
@@ -10,6 +10,7 @@ import (
 	"runtime"
 	"time"
 
+	"github.com/grafana/agent/internal/useragent"
 	"github.com/prometheus/common/version"
 )
 
@@ -27,6 +28,7 @@ type Report struct {
 	Metrics      map[string]interface{} `json:"metrics"`
 	Os           string                 `json:"os"`
 	Arch         string                 `json:"arch"`
+	DeployMode   string                 `json:"deployMode"`
 }
 
 func sendReport(ctx context.Context, seed *AgentSeed, interval time.Time, metrics map[string]interface{}) error {
@@ -38,6 +40,7 @@ func sendReport(ctx context.Context, seed *AgentSeed, interval time.Time, metric
 		Arch:         runtime.GOARCH,
 		Interval:     interval,
 		Metrics:      metrics,
+		DeployMode:   useragent.GetDeployMode(),
 	}
 	out, err := json.MarshalIndent(report, "", " ")
 	if err != nil {